Project_Documents/Archive/Calendar/Modernization/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
alembic==1.7.1
pydantic==1.8.2
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.19.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
email-validator==1.1.3
pytz==2021.1
aiohttp==3.7.4
python-dotenv==0.19.0
psycopg2-binary==2.9.1
